Add new test case
